ci:remove playwright relevant changes. Change 'linting' to 'formatting'
