+++ .gitignore
@@ -30,7 +30,3 @@ npm-debug.log*
 /platforms
 /plugins
 /www
-/test-results/
-/playwright-report/
-/blob-report/
-/playwright/.cache/
