+++ package.json
@@ -7,8 +7,8 @@
     "build": "vue-cli-service build",
     "test:unit": "vue-cli-service test:unit",
     "test:e2e": "vue-cli-service test:e2e",
-    "lint": "prettier --check .",
-    "lint-fix": "prettier --write ."
+    "format": "prettier --check .",
+    "format-fix": "prettier --write ."
   },
   "dependencies": {
     "@capacitor/app": "5.0.7",
@@ -25,7 +25,6 @@
   },
   "devDependencies": {
     "@capacitor/cli": "5.7.2",
-    "@playwright/test": "^1.43.1",
     "@types/jest": "^27.0.2",
     "@types/node": "^20.12.7",
     "@typescript-eslint/eslint-plugin": "^5.6.0",
@@ -55,5 +54,5 @@
   "engines": {
     "node": "20"
   },
-  "description": "An Ionic project"
+  "description": "An Ionic project template"
 }
